move classmethods to functions
